Major Changes.
I changed the structure of the previous code. This now follows the MVC
design Pattern.

This now has model, controller, view(empty), adapter(interface), and
main(main class) packages.

I have run this program, and it runs. All we need to do is implement our
assigned classes.

I made the Board class Singleton.

I removed all the methods that contains GUI components. We can implement
them in the view later in the course.


1. added a nested class called Tile in the Board class.

2. added alien in the Tile class, you can add the player there.

3. createAliens() ~ call to create aliens and placed them on the board.

4. aliensMove() ~ call to move aliens on the board [I'm currently
working on this method]

I added "InterruptedException" on some of the method because I'm testing
something. Feel free to remove them.

I'll be working on the Bomb class after I'm done with the Alien.

You can add the Shot class, I did not add it because you may declare it
inside the player class.


So far I've successfully printed the aliens on the compiler, for now
there image is "*" since we are not required to use any Swing
components.

I'm currently working on moving the aliens downward in runtime.